Remove deleted files from .vcxproj
Sometimes when you delete files, Visual Studio does not immediately save the changes to .vcxproj
This leads to git not including the differences made to .vcxproj, so when someone pulls your
commits, their .vcxproj will tell them to compile the deleted files.

To get around this, press "Save All" in Visual Studio.